db-facilities/pyproject.toml
[tool.poetry]
name = "facdb"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^1.2.3"
SQLAlchemy = "^1.4.3"
usaddress = "^0.5.10"
psycopg2-binary = "^2.8.6"
python-geosupport = { git = "https://github.com/NYCPlanning/python-geosupport.git", rev = "30a2097" }
python-dotenv = "^0.16.0"
typer = "^0.3.2"
tqdm = "^4.59.0"
PyYAML = "^5.4.1"
ipython = "^8.2.0"
boto3 = "^1.23.1"
python-crfsuite = { git = "https://github.com/garfieldnate/python-crfsuite.git", branch = "fix-python-3.10" }
pythainlp = "^2.3.2"

[tool.poetry.dev-dependencies]
pre-commit = "^2.11.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
facdb = "facdb.cli:app"
